requirements.txt
numpy==1.20.2
Pillow==8.2.0
tqdm==4.60.0
